maintenance: Skip setAgentAndTriggers for DB_NONE maintenance tasks
Maintenance scripts that are meant to operate without a database (those
that implement `getDbType()` and return `Maintenance::DB_NONE`) should
not call `setAgentAndTriggers` which results in allocation of a new DB
load balancer. Unconditionally calling this method makes it impossible
for scripts to run without a database or database configuration.

Bug: T260827
Bug: T237148
Bug: T238436
Change-Id: Ic69ea6f6ebda1ca6be5cbbd96c7a4a48adce34bd